Move some functions from Store to specific widgets
